Phase-III/backend/pyproject.toml
# [Task]: T002, T005, [From]: specs/001-task-crud-api/spec.md#Requirements
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "task-crud-api"
version = "0.1.0"
description = "Task CRUD API for Evolution of Todo - Phase 2"
requires-python = ">=3.11"
authors = [{name = "Development Team"}]
readme = "README.md"

dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlmodel==0.0.14",
    "sqlalchemy[asyncio]==2.0.23",
    "asyncpg==0.29.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "cryptography==41.0.7",
    "python-multipart==0.0.6",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "httpx==0.25.1",
]

[project.optional-dependencies]
dev = [
    "ruff==0.1.8",
    "black==23.12.0",
    "mypy==1.7.1",
]

[tool.setuptools]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [".git", "__pycache__", ".venv", "venv", ".pytest_cache"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "RUF"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--strict-markers"
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "contract: contract tests",
]
